Create n_body_simulation
This project is a discrete simulation solution of an N-body gravitational force problem. Given some number of bodies in empty space, the bodies are allowed to move freely and interact with each other under gravitational force, although in this simulation you will not simulate collisions between the bodies. The result yields an interesting animation.