Create Error_Detection_Homework
In this project you will simulate the transmission of data over a noisy communications channel (like a network cable or a wi-fi connection or a cell phone channel), and the subsequent reception of the data, and finally detection of any errors that occurred during transmission.  The user is expected to supply one character. The program is supposed to display the binary representation of that character and then the binary representation with a parity bit added on along with other variations of that. The program is supposed to show how many bits were changed throughout the program.