Create Error_Correction_HW
In this project you will use arrays of numbers to create groups of bits (binary digits) that are to be sent over a simulated communications channel from one computer to another. The sender uses parity bits to encode extra information in the data that allows the receiver to detect and correct errors in the transmitted bits. The simulated communications channel will be a function that, with low probability, flips bits randomly. The user is expected to enter a string. The string is then converted into bits and transmitted. Then the code checks to see if the string was transmitted without any errors. If there were errors, the program tries to correct them. At the end, the program either prints the string with errors that could not be fixed, or a corrected string. The constants used in the program are the fill character, the desired parity and the error probability.